lambda/requirements.txt
line-bot-sdk>=3.11.0
boto3>=1.35.0
fastapi>=0.115.0
uvicorn>=0.32.0
